move sim data out of src
